requirements.txt
dj-database-url==0.5.0
Django==2.2.24
django-heroku==0.0.0
djangorestframework==3.13.1
gunicorn==20.1.0

psycopg2-binary==2.9.2
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
